add type guard output for plain structs
This also will currently output partial checkers for applied generics.

I'm not sure what to do about structures/unions with open names,
but I think it might be useful to just do partial checkers on those too,
which means we'll at the very least get checkers for concrete parts and
it'll be more uniform (in that all structures will have some level of
type guard output).