add embedded union as definition
This is meant to be a union that dictates that the type tag exists in
the payload instead of outside. Instead of `tag + data`
we get only the payload, but with the type tag inside of it.

Since this requires a struct of some kind, the definition itself has a
data dependency on a struct, so we can ensure that this exists when
generating code for it.