check references for UntaggedUnion & AppliedName
This should finish up the reference check PR and make it so that
we can't reference unknown types.

This also meant restructuring how open names were identified, which I
think was an upside.

There is some dynamicity with regards to recursive references that I've
created a `LooseReference` case for. I'm not sure that it's something
that will last. I might rework it completely or harden it a bit.

One idea for `LooseReference` is to only allow it for pointer types,
which would bring it more in line with reasonable structure for all
languages. If a language doesn't care about pointers it'll simply have
the output omit the pointer, but they'll still have to work with that
behind the scenes.